Error functions

The group of loss functions used to calculate OOB errors.

error_funcs(ytrue, ypred, wcase = 1, method)

dev_poiss(ytrue, ypred, wcase = 1)

dev_gamma(ytrue, ypred, wcase = 1)

wgt_mse(ytrue, ypred, wcase = 1)

mcc(ytrue, ypred, wcase = 1)

Arguments

ytrue
numeric vector of the true target values.

ypred
numeric vector of the predicted target values.

wcase
numeric vector of case weights.

method
string specifying the type of random forest. One of the following: 'anova', 'class', 'poisson', 'gamma' or 'lognormal'.

Functions

dev_poiss: Poisson deviance loss function for Poisson regression

dev_gamma: Gamma deviance loss function for gamma regression

wgt_mse: Weighted mean squared error loss function for anova and lognormal regression

mcc: Matthews correlation coefficient for binary classification