Predictions from a Fitted Rpart Object

Returns a vector of predicted responses from a fitted rpart object.

# S3 method for rpart
predict(object, newdata,
       type = c("vector", "prob", "class", "matrix"),
       na.action = na.pass, ...)

Arguments

object
fitted model object of class "rpart". This is assumed to be the result of some function that produces an object with the same named components as that returned by the rpart function.

newdata
data frame containing the values at which predictions are required. The predictors referred to in the right side of formula(object) must be present by name in newdata. If missing, the fitted values are returned.

type
character string denoting the type of predicted value returned. If the rpart object is a classification tree, then the default is to return prob predictions, a matrix whose columns are the probability of the first, second, etc. class. (This agrees with the default behavior of tree). Otherwise, a vector result is returned.

na.action
a function to determine what should be done with missing values in newdata. The default is to pass them down the tree using surrogates in the way selected when the model was built. Other possibilities are na.omit and na.fail.

...
further arguments passed to or from other methods.

Value

A new object is obtained by dropping newdata down the object. For factor predictors, if an observation contains a level not used to grow the tree, it is left at the deepest possible node and frame$yval at the node is the prediction.

If type = "vector":
vector of predicted responses. For regression trees this is the mean response at the node, for Poisson trees it is the estimated response rate, and for classification trees it is the predicted class (as a number).

If type = "prob":
(for a classification tree) a matrix of class probabilities.

If type = "matrix":
a matrix of the full responses (frame$yval2 if this exists, otherwise frame$yval). For regression trees, this is the mean response, for Poisson trees it is the response rate and the number of events at that node in the fitted tree, and for classification trees it is the concatenation of at least the predicted class, the class counts at that node in the fitted tree, and the class probabilities (some versions of rpart may contain further columns).

If type = "class":
(for a classification tree) a factor of classifications based on the responses.

Details

This function is a method for the generic function predict for class "rpart". It can be invoked by calling predict for an object of the appropriate class, or directly by calling predict.rpart regardless of the class of the object.

See also

Examples

z.auto <- rpart(Mileage ~ Weight, car.test.frame)

#> Error in is.data.frame(data): object 'car.test.frame' not found

predict(z.auto)

#> Error in predict(z.auto): object 'z.auto' not found

fit <- rpart(Kyphosis ~ Age + Number + Start, data = kyphosis)

#> Error in is.data.frame(data): object 'kyphosis' not found

predict(fit, type = "prob")   # class probabilities (default)

#> Error in predict(fit, type = "prob"): object 'fit' not found

predict(fit, type = "vector") # level numbers

#> Error in predict(fit, type = "vector"): object 'fit' not found

predict(fit, type = "class")  # factor

#> Error in predict(fit, type = "class"): object 'fit' not found

predict(fit, type = "matrix") # level number, class frequencies, probabilities

#> Error in predict(fit, type = "matrix"): object 'fit' not found

sub <- c(sample(1:50, 25), sample(51:100, 25), sample(101:150, 25))
fit <- rpart(Species ~ ., data = iris, subset = sub)
fit

#> n= 75 
#> 
#> node), split, n, loss, yval, (yprob)
#>       * denotes terminal node
#> 
#> 1) root 75 50 setosa (0.3333333 0.3333333 0.3333333)  
#>   2) Petal.Width< 0.8 25  0 setosa (1.0000000 0.0000000 0.0000000) *
#>   3) Petal.Width>=0.8 50 25 versicolor (0.0000000 0.5000000 0.5000000)  
#>     6) Petal.Length< 4.8 25  1 versicolor (0.0000000 0.9600000 0.0400000) *
#>     7) Petal.Length>=4.8 25  1 virginica (0.0000000 0.0400000 0.9600000) *

table(predict(fit, iris[-sub,], type = "class"), iris[-sub, "Species"])

#>             
#>              setosa versicolor virginica
#>   setosa         25          0         0
#>   versicolor      0         22         2
#>   virginica       0          3        23